% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{dublinbikes}
\alias{dublinbikes}
\alias{dublinbikes-package}
\title{Dublin bikes 2020 Q1 usage data.}
\format{
A data frame with 2,228,278 rows (Observations) with 11 columns(variables)
}
\source{
\url{https://data.smartdublin.ie}
}
\usage{
dublinbikes
}
\description{
Quarterly rollups of historic bike data.
Static data provides information like bike station position, number of bike stands, payment terminal availability
Formatted to be convenient for data analysis with R.
}
\keyword{datasets}
